% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/thornthwaite_mather__calc_accumulated_potential_water_loss.R
\name{calc_APWL}
\alias{calc_APWL}
\title{Calculate accumulated potential water loss.}
\usage{
calc_APWL(max_soil_moisture_in, soil_moisture_in)
}
\arguments{
\item{max_soil_moisture_in}{Maximum soil moisture storage with the root zone, in inches.}

\item{soil_moisture_in}{Soil moisture, in inches.}
}
\value{
Accumulated potential water loss in inches.
}
\description{
This function returns a calculated accumulated potential water loss value given
the current soil moisture and maximum soil moisture for the soil.
}
\details{
The value is generated from a nonlinear equation fit to the data values given in
Thornthwaite and Mather (1957).
}
\references{
Thornthwaite, C.W., and Mather, J.R., 1957, Instructions and tables for
computing potential evapotranspiration and the water balance: Publications
in Climatology, v. 10, no. 3, p. 1–104.
}
