% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeus__daylight_hours.R
\name{calc_daylight_hours}
\alias{calc_daylight_hours}
\title{Calculate number of daylight hours}
\usage{
calc_daylight_hours(month, day, year, latitude)
}
\arguments{
\item{month}{Month of year (1-12)}

\item{day}{Day of month (1-31)}

\item{year}{Year (YYYY)}

\item{latitude}{Latitude of station, in decimal degrees?}
}
\value{
Number of daylight hours
}
\description{
Returns the number of daylight hours for a given latitude and time of year.
}
\references{
Meeus, J.H., 1991, Astronomical algorithms: Willmann-Bell, Incorporated, 477 p.
}
