% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite_mather__calc_soil_moisture.R
\name{calc_soil_moisture}
\alias{calc_soil_moisture}
\title{Calculate soil moisture.}
\usage{
calc_soil_moisture(max_soil_moisture_in, apwl_in)
}
\arguments{
\item{max_soil_moisture_in}{Maximum soil moisture storage with the root zone, in inches.}

\item{apwl_in}{Accumulated potential water loss, in inches.}
}
\value{
Soil moisture in inches.
}
\description{
This function returns a calculated soil moisture value given the current
accumulated potential water loss (APWL) and maximum soil moisture for the soil.
}
\details{
Accumulated potential water loss is calculated separately and represents the
sum of the difference between the potential evapotranspiration and the
actual evapotranspiration.

The value is generated from a nonlinear equation fit to the data values given in
Thornthwaite and Mather (1957).
}
\references{
Thornthwaite, C.W., and Mather, J.R., 1957, Instructions and tables for
computing potential evapotranspiration and the water balance: Publications
in Climatology, v. 10, no. 3, p. 1–104.
}
