% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeus__calc_solar_declination.R
\name{calc_solar_declination}
\alias{calc_solar_declination}
\title{Calculate solar declination.}
\usage{
calc_solar_declination(fractional_year_radians)
}
\arguments{
\item{fractional_year_radians}{Day of year, expressed as fraction of year in radians.}
}
\value{
Solar declination, in radians.
}
\description{
Calculate solar declination.
}
\references{
Meeus, J.H., 1991, Astronomical algorithms: Willmann-Bell, Incorporated, 477 p.
}
