% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite_mather__a_coefficient.R
\name{calc_TM_a}
\alias{calc_TM_a}
\title{Calculate exponent 'a' in Thornthwaite's calculation of potential evapotranspiration.}
\usage{
calc_TM_a(heat_index_I)
}
\arguments{
\item{heat_index_I}{Annual heat index at a climate station.}
}
\value{
Exponent 'a' in Thornthwaite's PET calculation method.
}
\description{
This function returns the exponent 'a' in Thornthwaite's potential evapotranspiration
calculation method. The version implemented here is given as equation 9 in Thornthwaite (1948).
}
\references{
Thornthwaite, C.W., 1948, An Approach toward a Rational Classification of
 Climate: Geographical Review, v. 38, no. 1, p. 55.
}
