% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite_mather__i_monthly_heat_index.R
\name{calc_TM_i}
\alias{calc_TM_i}
\title{Calculate monthly heat index \code{i} in Thornthwaite's calculation of potential evapotranspiration.}
\usage{
calc_TM_i(air_temp_C)
}
\arguments{
\item{air_temp_C}{Mean monthly air temperature at a station, in degrees Celcius.}
}
\value{
Monthly heat index \code{i} in Thornthwaite's PET calculation method.
}
\description{
This function returns monthly heat index \code{i} used in Thornthwaite's potential
evapotranspiration calculation method. The equation used here is given in the text
of Appendix 1 (page 89) in Thornthwaite (1948).
}
\references{
Thornthwaite, C.W., 1948, An Approach toward a Rational Classification of
 Climate: Geographical Review, v. 38, no. 1, p. 55.
}
