% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite_mather__PE.R
\name{calc_TM_PET}
\alias{calc_TM_PET}
\title{Calculate Thornthwaite's potential evapotranspiration for normal temperature conditions.}
\usage{
calc_TM_PET(air_temp_C, heat_index_I, TM_a)
}
\arguments{
\item{air_temp_C}{Mean monthly air temperature at a station, in degrees Celcius.}

\item{heat_index_I}{Annual heat index \code{I} for a station.}

\item{TM_a}{Exponent \code{a} in Thornthwaite's PET calculation.}
}
\value{
Unadjusted potential evapotranspiration in millimeters.
}
\description{
This function returns a value for potential evapotranspiration as calculated by
Thornthwaite's potential evapotranspiration calculation method. This function is designed to
be applied when air temperatures fall below 26.5 degrees C. The equation used here
is given as equation 10 in Appendix 1 of Thornthwaite (1948). Note that the version
implemented here returns *millimeters*, rather than *centimeters* of PET, and therefore
the coefficient used here is '16' rather than '1.6'.
}
\references{
Thornthwaite, C.W., 1948, An Approach toward a Rational Classification of
 Climate: Geographical Review, v. 38, no. 1, p. 55.
}
