% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thornthwaite_mather__PE_high_temperature.R
\name{calc_TM_PET_hi_temp}
\alias{calc_TM_PET_hi_temp}
\title{Calculate Thornthwaite's potential evapotranspiration for high air temperature conditions.}
\usage{
calc_TM_PET_hi_temp(air_temp_C)
}
\arguments{
\item{air_temp_C}{Mean monthly air temperature at a station, in degrees Celcius.}
}
\value{
Unadjusted potential evapotranspiration in millimeters.
}
\description{
This function returns a value for potential evapotranspiration as calculated by
Thornthwaite's potential evapotranspiration calculation method. This function is designed to
be applied when air temperatures fall between 26.5 and 38 degrees C. The data used here
are given in figure 13 of Appendix 1 in Thornthwaite (1948).
}
\references{
Thornthwaite, C.W., 1948, An Approach toward a Rational Classification of
 Climate: Geographical Review, v. 38, no. 1, p. 55.
}
